Implement map abstraction
Implement a LevelMap class that is an abstraction of an level map.
This class loads a map configuration from a file. In addition, new
resources were included to test this new implementation.

Also, implement a IntroState to test the new LevelMap class and to
try the resources given by the game designers.