Bug fixing.
Fixed the bug where the player takes too much damage when walking
on the mekabug while colliding. Also fixed the balls damage and
removed their hitbox, so the player doenst get stuck when they
are attacking. Was implemented the damage for the samurai and
decker class, but it still needs to fix the damage dealt.