Changed the hitbox in AttackClass to attackHitbox.
Before, the AttackClass was using the hitbox variable to check if
ocurred collision between the attack and another game object, wich
made the balls imune to the attack damage. Now, the AttackClass
uses the attackHitbox to see if ocurred collision, wich makes
all game objects collide with the attack, if they can be attacked.